+++ web/embed.go
@@ -3,5 +3,5 @@ package web
 
 import "embed"
 
-//go:embed index.html
+//go:embed index.html css js
 var FS embed.FS